Projet OPA/Data et Code/Partie 2/requirements.txt
pandas==2.0.3
psycopg2-binary==2.9.7
python-decouple==3.8
python_dateutil==2.8.2
Requests==2.31.0
tqdm==4.66.1
